# Fix Symbol Equals emtpy